Move wrapper package back to SDK
TF uses this, so SDK is a decent place for it. It doesn't create
dependencies between it and API.
