+++ api/go.mod
@@ -12,6 +12,7 @@ require (
 	github.com/hashicorp/go-uuid v1.0.2
 	github.com/mr-tron/base58 v1.2.0
 	github.com/stretchr/testify v1.7.0
+	golang.org/x/crypto v0.0.0-20200820211705-5c72a883971a // indirect
 	golang.org/x/time v0.0.0-20200630173020-3af7569d3a1e
 	google.golang.org/grpc v1.35.0
 	google.golang.org/protobuf v1.27.1
