+++ internal/cmd/commands/config/encryptdecrypt.go
@@ -8,7 +8,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/boundary/internal/cmd/base"
-	"github.com/hashicorp/boundary/internal/cmd/base/wrapper"
+	"github.com/hashicorp/boundary/sdk/wrapper"
 	"github.com/hashicorp/go-secure-stdlib/configutil"
 	"github.com/mitchellh/cli"
 	"github.com/posener/complete"
