+++ sdk/go.mod
@@ -3,10 +3,9 @@ module github.com/hashicorp/boundary/sdk
 go 1.15
 
 require (
-	github.com/davecgh/go-spew v1.1.1 // indirect
-	github.com/kr/pretty v0.2.0 // indirect
-	github.com/kr/text v0.2.0 // indirect
+	github.com/hashicorp/go-kms-wrapping v0.6.6
+	github.com/hashicorp/go-secure-stdlib/configutil v0.1.2
+	github.com/hashicorp/go-secure-stdlib/strutil v0.1.1
 	github.com/stretchr/testify v1.7.0
 	google.golang.org/protobuf v1.27.1
-	gopkg.in/check.v1 v1.0.0-20190902080502-41f04d3bba15 // indirect
 )
